Rename Servers -> Server && Refactor "isNewer" garbage as stop gap before full removal
