--- src/main/java/metadev/digital/metacustomitemslib/server/Servers.java
@@ -1,260 +1,0 @@
-package metadev.digital.metacustomitemslib.server;
-
-import org.bukkit.Bukkit;
-
-public class Servers {
-
-	// *******************************************************************
-	// Version detection TODO: Unfuck this, this is a terrible way to do this
-	// *******************************************************************
-
-	@Deprecated
-	public static boolean isMC121() {
-		return Bukkit.getBukkitVersion().contains("1.21");
-	}
-
-	@Deprecated
-	public static boolean isMC120() {
-		return Bukkit.getBukkitVersion().contains("1.20");
-	}
-
-	@Deprecated
-	public static boolean isMC119() {
-		return Bukkit.getBukkitVersion().contains("1.19");
-	}
-
-	@Deprecated
-	public static boolean isMC118() {
-		return Bukkit.getBukkitVersion().contains("1.18");
-	}
-
-	@Deprecated
-	public static boolean isMC117() {
-		return Bukkit.getBukkitVersion().contains("1.17");
-	}
-
-	@Deprecated
-	public static boolean isMC1162() {
-		return Bukkit.getBukkitVersion().contains("1.16.2");
-	}
-
-	@Deprecated
-	public static boolean isMC116() {
-		return Bukkit.getBukkitVersion().contains("1.16");
-	}
-
-	@Deprecated
-	public static boolean isMC115() {
-		return Bukkit.getBukkitVersion().contains("1.15");
-	}
-
-	@Deprecated
-	public static boolean isMC114() {
-		return Bukkit.getBukkitVersion().contains("1.14");
-	}
-
-	@Deprecated
-	public static boolean isMC113() {
-		return Bukkit.getBukkitVersion().contains("1.13");
-	}
-
-	@Deprecated
-	public static boolean isMC112() {
-		return Bukkit.getBukkitVersion().contains("1.12");
-	}
-
-	@Deprecated
-	public static boolean isMC111() {
-		return Bukkit.getBukkitVersion().contains("1.11");
-	}
-
-	@Deprecated
-	public static boolean isMC110() {
-		return Bukkit.getBukkitVersion().contains("1.10");
-	}
-
-	@Deprecated
-	public static boolean isMC19() {
-		return Bukkit.getBukkitVersion().contains("1.9");
-	}
-
-	@Deprecated
-	public static boolean isMC18() {
-		return Bukkit.getBukkitVersion().contains("1.8");
-	}
-
-	@Deprecated
-	public static boolean isMC121OrNewer() {
-		if (isMC121())
-			return true;
-		else if (isMC120() || isMC119() || isMC118() || isMC117() || isMC1162() || isMC116() || isMC115() || isMC114() || isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC120OrNewer() {
-		if (isMC120())
-			return true;
-		else if (isMC119() || isMC118() || isMC117() || isMC1162() || isMC116() || isMC115() || isMC114() || isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC119OrNewer() {
-		if (isMC119())
-			return true;
-		else if (isMC118() || isMC117() || isMC1162() || isMC116() || isMC115() || isMC114() || isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC118OrNewer() {
-		if (isMC118())
-			return true;
-		else if (isMC117() || isMC1162() || isMC116() || isMC115() || isMC114() || isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC117OrNewer() {
-		if (isMC117())
-			return true;
-		else if (isMC1162() || isMC116() || isMC115() || isMC114() || isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC1162OrNewer() {
-		if (isMC1162())
-			return true;
-		else if (isMC116() || isMC115() || isMC114() || isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC116OrNewer() {
-		if (isMC116())
-			return true;
-		else if (isMC115() || isMC114() || isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC115OrNewer() {
-		if (isMC115())
-			return true;
-		else if (isMC114() || isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC114OrNewer() {
-		if (isMC114())
-			return true;
-		else if (isMC113() || isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC113OrNewer() {
-		if (isMC113())
-			return true;
-		else if (isMC112() || isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC112OrNewer() {
-		if (isMC112())
-			return true;
-		else if (isMC111() || isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC111OrNewer() {
-		if (isMC111())
-			return true;
-		else if (isMC110() || isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC110OrNewer() {
-		if (isMC110())
-			return true;
-		else if (isMC19() || isMC18())
-			return false;
-		return true;
-	}
-
-	@Deprecated
-	public static boolean isMC19OrNewer() {
-		if (isMC19())
-			return true;
-		else if (isMC18())
-			return false;
-		return true;
-	}
-
-	// *******************************************************************
-	// Version detection
-	// *******************************************************************
-	public static String getServerVersion() {
-		return Bukkit.getBukkitVersion().split("-")[0];
-	}
-
-	public static boolean serverIsAfterProvidedVersion(String compareVersion) {
-		return getServerVersion().compareTo(compareVersion) >= 0;
-	}
-
-	public static ArchitectureEnum getServerArchitecture() {
-		String serverName = Bukkit.getServer().getName();
-		String versionTitle = Bukkit.getServer().getVersion().toLowerCase();
-
-		if (serverName.equalsIgnoreCase("Paper") && versionTitle.contains("paper")) return ArchitectureEnum.PAPER;
-		if (serverName.equalsIgnoreCase("Purpur") && versionTitle.contains("purpur")) return ArchitectureEnum.PURPUR;
-		if (serverName.equalsIgnoreCase("CraftBukkit") && versionTitle.contains("spigot")) return ArchitectureEnum.SPIGOT;
-		if (serverName.equalsIgnoreCase("CraftBukkit") && versionTitle.contains("bukkit")) return ArchitectureEnum.BUKKIT;
-		if (serverName.equalsIgnoreCase("Glowstone")) return  ArchitectureEnum.GLOWSTONE;
-
-		return ArchitectureEnum.UNKNOWN;
-	}
-
-	public static boolean isGlowstoneServer() {
-		return Bukkit.getServer().getName().equalsIgnoreCase("Glowstone");
-	}
-
-    public static boolean isPaperServer() {
-        return Bukkit.getServer().getName().equalsIgnoreCase("Paper")
-                && Bukkit.getServer().getVersion().toLowerCase().contains("paper");
-    }
-
-    public static boolean isPurpurServer() {
-        return Bukkit.getServer().getName().equalsIgnoreCase("Purpur")
-                && Bukkit.getServer().getVersion().toLowerCase().contains("purpur");
-    }
-
-	public static boolean isSpigotServer() {
-		return Bukkit.getServer().getName().equalsIgnoreCase("CraftBukkit")
-				&& Bukkit.getServer().getVersion().toLowerCase().contains("spigot");
-	}
-
-	public static boolean isCraftBukkitServer() {
-		return Bukkit.getServer().getName().equalsIgnoreCase("CraftBukkit")
-				&& Bukkit.getServer().getVersion().toLowerCase().contains("bukkit");
-	}
-
-}
